requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
babel==2.16.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
cssselect2==0.7.0
distlib==0.3.9
Django==5.2.5
django-environ==0.11.2
django-extra-views==0.14.0
django-haystack==3.3.0
django-oscar==3.2.5
django-paypal==2.1
django-phonenumber-field==6.4.0
django-tables2==2.3.4
django-treebeard==4.7.1
django-widget-tweaks==1.4.8
djangorestframework==3.16.0
environ==1.0
factory-boy==3.2.1
Faker==33.3.1
filelock==3.16.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
legacy-cgi==2.6.3
lxml==5.3.0
oscrypto==1.3.0
packaging==24.2
phonenumbers==8.13.52
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
purl==1.6
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.1.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
razorpay==1.4.2
reportlab==4.2.5
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.3.1
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
virtualenv==20.28.1
webencodings==0.5.1
xhtml2pdf==0.2.16
